Add files via upload
When the data suggest a curved or quadratic pattern, a quadratic line of best fit, also
known as a quadratic regression, is used in statistics to model the relationship between two
variables. A quadratic line of best fit is a curve that minimizes the sum of the squared
difference between the observed and predicted values in the context of a scatter plot, where
data points are plotted. An example of its use in our everyday life is when there is noticeable
curvature or seasonality in the data, a quadratic line of best fit can be useful for modeling
temperature patterns over time.

When there is a linear or approximately linear trend in the data, a linear line of best
fit, which is commonly used in linear regression, is used to model the relationship between
two variables. It is a straight line whose slope minimizes the sum of the squared differences
between observed and predicted values. Linear regression is a popular statistical technique
with numerous applications and the linear line of best fit serves several functions. Sport
performance, education, Sale and Marketing all are everyday things that we do that require
the line of best fit to see trends based on two variables